Starting making Jacobian work.
