+++ Project.toml
@@ -6,7 +6,6 @@ version = "0.1.1"
 [deps]
 AbstractNeuralNetworks = "60874f82-5ada-4c70-bd1c-fa6be7711c8a"
 ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
-GeometricMachineLearning = "194d25b2-d3f5-49f0-af24-c124f4aa80cc"
 KernelAbstractions = "63c18a36-062a-441e-b654-da1e3ab1ce7c"
 Latexify = "23fbe1c1-3f47-55db-b15f-69d7ec21a316"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
@@ -17,7 +16,6 @@ Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [compat]
 AbstractNeuralNetworks = "0.1, 0.3"
-GeometricMachineLearning = "0.3.5"
 KernelAbstractions = "0.9"
 RuntimeGeneratedFunctions = "0.5"
 SafeTestsets = "0.1"
